requirements.txt
flask
pandas
openpyxl
python-docx
PyMuPDF